requirements.txt
Flask
python-dotenv
requests
werkzeug
pdfminer.six
python-docx
numpy
reportlab